Create prototype mastermind
I wrote a basic version of the game mastermind using numbers instead of
colors.

I created a class Game, which creates a 4 digit number with each digit
between 1 and 6 upon initialization and stores it in a class variable
It also contains class variables for the player choice, hints, number of
rounds to play and a boolean value whether the game is still running

I wrote a method to ask for a 4 digit user input, which is stored in the
class variable

Another method compares the user input to the code and adds an X to the
hints if the correct number is in the correct position and an O if a
number the player input contains is included in the code.
Still trying to figure out how to make this work properly, as I'm
getting more Os than I should in case of number duplicates.

The final method so far is startig the game, and looping as long as the
number of turns is not over and the code has not been cracked