Create working check algorithm
I updated the check algorithm so that it now returns the correct hints.

I did this by first creating copies of the player guess and code array.
Then the algorithm compares each item in the copied player guess array
to the copied code array by index (code [0] = guess[=]). If the item is
equal, an "X" is added to the hints and the item is removed from either
array.

After that, I checked for each item from the guess array if it is
included in the code. If so, an "O" is added to the guesses and the item
is removed from the code array. so that a duplicate in the remaining
guesses array does not give another hint in case there is no duplicate
in the code array