Create computer guesser
I added a prompt whether the player wants to crack randomly generated
code or wants to set a code for the computer to crack.

In case the player sets the code, an array with all possible codes is
generated. The computer then randomly picks one of those possibilities
and it's then compared to the player code, returning the same hints the
player would get when trying to crack the code.

So far I only implemented a check depending on the number of exact
matches ("X" hints).
For each remaining possible code, a check is made whether the code would
be a valid choice cosidering the possibilities depending on the "X"s
e.g. player code = 1234, computer guess = 1111 => one "X"
In this case, a valid code has to look like one of these:
1XXX // X1XX // XX1X // XX1X
If the code does not look alike, it is removed from the array of
possibilities, otherwise the next possible code is checked.
Then another pick is made from the possibilities and checked and so on
until either the number of rounds is over or the code is cracked.